Refactor: 리스트 컴포넌트화
